feat: add mms only in the end + cleanup
